refactor: directory structure to be more
