+++ internal/message/handler.go
@@ -1,4 +1,4 @@
-package services
+package message
 
 import (
 	"errors"
