+++ internal/message/handler_integration_test.go
