--- types/interfaces.go
@@ -1,9 +1,0 @@
-package types
-
-type MessageStore interface {
-	GetMessage(id string) (Message, error)
-	GetMessages() ([]Message, error)
-	CreateMessage(message *Message) error
-	UpdateMessage(id string, message *Message) error
-	DeleteMessage(id string) error
-}
